+++ Changes.md
@@ -21,6 +21,7 @@ Fixes
 - Linux : Fixed startup from installation directory containing spaces.
 - SceneReader : Fixed loading of float-precision orientations from the USD `PointInstancer.orientationsf` attribute. Previously only the half-precision `orientations` attribute was supported.
 - Viewer : Fixed colourspace handling for PNG files loaded as textures for custom OpenGL shaders.
+- PointInstancer : Fixed potential interactive render update bug.
 
 API
 ---